add tiff/gif support, get ready to rip the bandaid off
