+++ cmd/exif/main.go
@@ -1,27 +1,21 @@
 package main
 
 import (
-	"bufio"
-	"bytes"
 	"context"
 	"flag"
 	"fmt"
 	"image"
-	"image/jpeg"
-	"image/png"
 	"io"
 	"log"
 	"log/slog"
 	"os"
 	"strconv"
 
+	"github.com/aaronland/go-image/decode"
+	"github.com/aaronland/go-image/encode"
 	"github.com/aaronland/go-image/rotate"
 	"github.com/dsoprea/go-exif/v3"
-	"github.com/dsoprea/go-heic-exif-extractor/v2"
-	"github.com/dsoprea/go-jpeg-image-structure/v2"
-	"github.com/dsoprea/go-png-image-structure/v2"
 	"github.com/gabriel-vasile/mimetype"
-	"github.com/strukturag/libheif-go"
 )
 
 func RotateFromOrientation(im image.Image, mtype *mimetype.MIME, ifd *exif.Ifd) (bool, image.Image, error) {
@@ -58,9 +52,6 @@ func RotateFromOrientation(im image.Image, mtype *mimetype.MIME, ifd *exif.Ifd) 
 		return false, im, nil
 	}
 
-	slog.Info("ROTATE", "orientation", orientation)
-	return true, im, nil
-
 	ctx := context.Background()
 	r_im, err := rotate.RotateImageWithOrientation(ctx, im, orientation)
 
@@ -69,141 +60,6 @@ func RotateFromOrientation(im image.Image, mtype *mimetype.MIME, ifd *exif.Ifd) 
 	}
 
 	return true, r_im, nil
-}
-
-func ImageFromHEIC(body []byte) (image.Image, error) {
-
-	// https://github.com/spacestation93/heif_howto
-
-	// First decode the HEIC image
-
-	im_ctx, err := libheif.NewContext()
-
-	if err != nil {
-		return nil, fmt.Errorf("Failed to create new libheif context, %w", err)
-	}
-
-	err = im_ctx.ReadFromMemory(body)
-
-	if err != nil {
-		return nil, fmt.Errorf("Failed to read input data, %w", err)
-	}
-
-	im_handle, err := im_ctx.GetPrimaryImageHandle()
-
-	if err != nil {
-		return nil, fmt.Errorf("Failed to derive primary image handler, %w", err)
-	}
-
-	h_im, err := im_handle.DecodeImage(libheif.ColorspaceUndefined, libheif.ChromaUndefined, nil)
-
-	if err != nil {
-		return nil, fmt.Errorf("Failed to decode image, %w", err)
-	}
-
-	im, err := h_im.GetImage()
-
-	if err != nil {
-		return nil, fmt.Errorf("Failed to create image.Image, %w", err)
-	}
-
-	return im, nil
-}
-
-func DecodeImage(body []byte) (image.Image, *mimetype.MIME, *exif.Ifd, error) {
-
-	var ifd *exif.Ifd
-	var im image.Image
-
-	br := bytes.NewReader(body)
-
-	im, im_fmt, err := image.Decode(br)
-
-	if err != nil {
-		// Check error here...
-		slog.Warn("Failed to decode image natively", "error", err)
-	}
-
-	mtype := mimetype.Detect(body)
-	slog.Info("m", "type", mtype)
-
-	switch im_fmt {
-	case "jpeg":
-
-		jmp := jpegstructure.NewJpegMediaParser()
-		mc, err := jmp.ParseBytes(body)
-
-		if err != nil {
-			return nil, nil, nil, err
-		}
-
-		jpg_ifd, _, err := mc.Exif()
-
-		if err != nil {
-			slog.Warn("Failed to derive EXIF", "error", err)
-		} else {
-			ifd = jpg_ifd
-		}
-
-	case "png":
-
-		mp := pngstructure.NewPngMediaParser()
-
-		mc, err := mp.ParseBytes(body)
-
-		if err != nil {
-			return nil, nil, nil, err
-		}
-
-		png_ifd, _, err := mc.Exif()
-
-		if err != nil {
-			slog.Warn("Failed to derive EXIF", "error", err)
-		} else {
-			ifd = png_ifd
-		}
-
-	default:
-
-		switch mtype.String() {
-		case "image/heic":
-
-			heic_im, err := ImageFromHEIC(body)
-
-			if err != nil {
-				return nil, nil, nil, err
-			}
-
-			im = heic_im
-
-			mp := heicexif.NewHeicExifMediaParser()
-			mc, err := mp.ParseBytes(body)
-
-			if err != nil {
-				return nil, nil, nil, err
-			}
-
-			heic_ifd, _, err := mc.Exif()
-
-			if err != nil {
-				slog.Warn("Failed to derive EXIF", "error", err)
-			} else {
-				ifd = heic_ifd
-			}
-
-			// Note: We are NOT removing or updating the Orientation tag
-			// (which is assigned but incorrect) in libheif because I can
-			// not figure out hwo to do that using the dsoprea packages
-			// without causing everything to panic later in the code. Instead
-			// we are accounting for this in RotateFromOrientation
-			// https://github.com/strukturag/libheif/issues/227
-
-		default:
-			return nil, nil, nil, fmt.Errorf("Unsupported media type")
-		}
-	}
-
-	return im, mtype, ifd, nil
 }
 
 func main() {
@@ -228,7 +84,7 @@ func main() {
 
 		// Open with EXIF
 
-		im, mtype, ifd, err := DecodeImage(body)
+		im, mtype, ifd, err := decode.DecodeImage(body)
 
 		if err != nil {
 			log.Fatal(err)
@@ -257,7 +113,7 @@ func main() {
 			log.Fatal(err)
 		}
 
-		err = WriteJPEG(wr, im, ib, rotated)
+		err = encode.EncodeJPEG(wr, im, ib, nil)
 
 		err = wr.Close()
 
@@ -272,125 +128,31 @@ func main() {
 func newIfdBuilder(ifd *exif.Ifd, rotated bool) (*exif.IfdBuilder, error) {
 
 	var ib *exif.IfdBuilder
-	
+
 	if ifd != nil {
-		
+
 		ib = exif.NewIfdBuilderFromExistingChain(ifd)
-		
+
 		if rotated {
-			
+
 			ifdPath := "IFD0"
-			
+
 			ifd_ib, err := exif.GetOrCreateIbFromRootIb(ib, ifdPath)
-			
+
 			if err != nil {
 				return nil, err
 			}
-			
+
 			oint, _ := strconv.Atoi("1") // top left
 			oint16 := uint16(oint)
-			
+
 			err = ifd_ib.SetStandardWithName("Orientation", []uint16{oint16})
-			
+
 			if err != nil {
 				return nil, err
 			}
 		}
 	}
-	
+
 	return ib, nil
-}
-
-func WriteJPEG(wr io.Writer, im image.Image, ib *exif.IfdBuilder, rotated bool) error {
-
-	jpeg_opts := &jpeg.Options{
-		Quality: 100,
-	}
-
-	if ib == nil {
-		return jpeg.Encode(wr, im, jpeg_opts)
-	}
-
-	// Do EXIF dance
-
-	var im_buf bytes.Buffer
-	im_wr := bufio.NewWriter(&im_buf)
-
-	err := jpeg.Encode(im_wr, im, jpeg_opts)
-
-	if err != nil {
-		return err
-	}
-
-	im_wr.Flush()
-
-	// Write EXIF back to JPEG
-
-	jmp := jpegstructure.NewJpegMediaParser()
-
-	mp, err := jmp.ParseBytes(im_buf.Bytes())
-
-	if err != nil {
-		return err
-	}
-
-	sl := mp.(*jpegstructure.SegmentList)
-
-	err = sl.SetExif(ib)
-
-	if err != nil {
-		return err
-	}
-
-	err = sl.Write(wr)
-
-	if err != nil {
-		return err
-	}
-
-	return nil
-}
-
-func WritePNG(wr io.Writer, im image.Image, ib *exif.IfdBuilder, rotated bool) error {
-
-	if ib == nil {
-		return png.Encode(wr, im)
-	}
-
-	var im_buf bytes.Buffer
-	im_wr := bufio.NewWriter(&im_buf)
-
-	err := png.Encode(im_wr, im)
-
-	if err != nil {
-		return err
-	}
-
-	im_wr.Flush()
-
-	// Write EXIF back to PNG
-
-	png_parser := pngstructure.NewPngMediaParser()
-
-	mp, err := png_parser.ParseBytes(im_buf.Bytes())
-
-	if err != nil {
-		return err
-	}
-
-	cs := mp.(*pngstructure.ChunkSlice)
-
-	err = cs.SetExif(ib)
-
-	if err != nil {
-		return err
-	}
-
-	err = cs.WriteTo(wr)
-
-	if err != nil {
-		return err
-	}
-
-	return nil
 }
