+++ encode/heic.go
@@ -1,0 +1,3 @@
+package encode
+
+// https://github.com/strukturag/libheif-go/blob/master/encoder_test.go#L64
